make copy conditional

+++ operator/src/main/resources/scripts/livenessProbe.sh
@@ -17,32 +17,36 @@ function copySitCfgWhileRunning() {
   # This method is called while the server is already running, see 
   # 'copySitCfgWhileBooting' in 'startServer.sh' for a similar method that
   # is called just before the server boots.
+  #
+  # It will do nothing unless the environment variable DYNAMIC_CONFIG_OVERRIDE is set.
 
-  (
-  shopt -s nullglob  # force file matching 'glob' for loops below to run 0 times if 0 matches
-  local src_dir=${1?}
-  local tgt_dir=${2?}
-  local fil_prefix=${3?}
-  local local_fname
-  local tgt_file
-  trace "Copying files starting with '$src_dir/$fil_prefix' to '$tgt_dir' without the prefix."
-  mkdir -p $tgt_dir # TBD ignore any error?
-  for local_fname in {src_dir}/${fil_prefix}*.xml ; do
-    tgt_file=${local_fname/$fil_prefix//}   # strip out file prefix from source file
-    tgt_file=$(basename $tgt_file)          # strip out dir path since it's the source file path
-    tgt_file=$tgt_dir/$tgt_file             # add back in tgt dir path
-    [ -f "$tgt_file" ] && [ -z "$(diff $local_fname $tgt_file 2>&1)" ] && continue  # nothing changed
-    cp $local_fname $tgt_file # TBD ignore any error?
-    chmod 750 $tgt_file # TBD ignore any error?
-  done
-  for local_fname in {tgt_dir}/*.xml ; do
-    if [ -f "$src_dir/${fil_prefix}$(basename ${local_fname})" ]; then
-      continue
-    fi
-    trace "Deleting '$local_fname' since it has no corresponding '$src_dir' file."
-    rm -f $local_fname # TBD ignore any error?
-  done
-  )
+if [ ! "${DYNAMIC_CONFIG_OVERRIDE:-notset}" = notset ]; then
+    (
+    shopt -s nullglob  # force file matching 'glob' for loops below to run 0 times if 0 matches
+    local src_dir=${1?}
+    local tgt_dir=${2?}
+    local fil_prefix=${3?}
+    local local_fname
+    local tgt_file
+    trace "Copying files starting with '$src_dir/$fil_prefix' to '$tgt_dir' without the prefix."
+    mkdir -p $tgt_dir # TBD ignore any error?
+    for local_fname in {src_dir}/${fil_prefix}*.xml ; do
+      tgt_file=${local_fname/$fil_prefix//}   # strip out file prefix from source file
+      tgt_file=$(basename $tgt_file)          # strip out dir path since it's the source file path
+      tgt_file=$tgt_dir/$tgt_file             # add back in tgt dir path
+      [ -f "$tgt_file" ] && [ -z "$(diff $local_fname $tgt_file 2>&1)" ] && continue  # nothing changed
+      cp $local_fname $tgt_file # TBD ignore any error?
+      chmod 750 $tgt_file # TBD ignore any error?
+    done
+    for local_fname in {tgt_dir}/*.xml ; do
+      if [ -f "$src_dir/${fil_prefix}$(basename ${local_fname})" ]; then
+        continue
+      fi
+      trace "Deleting '$local_fname' since it has no corresponding '$src_dir' file."
+      rm -f $local_fname # TBD ignore any error?
+    done
+    )
+  fi
 }
